Some more localization. Turns out the actual PDF Renderer is still under development, so the most recent version was pulled and this library was updated to the most recent version: Updates on comments. PDFDecoder updated to support filters. TIFFPredictor added. Minor update to decryptors for use by decoders. Multiple classes updated to hand off rectangle creation to PDFile. TTFFont updated to support different Glyph types. Safety code, correction on ordering, small change to transforms in Type1Font. Actual implementation of functionality in FunctionType3. Correction to code that used XYRectFloat, which was updated to use doubles instead of floats. Addition of some new helper functions (lastIndexOf for Strings, full Vector comparisions) Minor change in determing step value in PatternType1. PDFFile has had many changes to support Xref types, seperation of file data and process data, handling of Strings, trailer and xref processing, file parsing, and (as stated previously) rectangle processing. Minor changes to PDFImage (now large changes did occur but PDFImage isn't exactly a port of the J2SE version of PDF Render so more complex work is needed to get fully up to date). Filter support in PDFObject. New manor of handling page sizes in PDFPage. Some new safety code in PDFParser. Minor bug fix in PDFRenderer. Change to PDFXref to handle different Xref types and extract the actual containment of Xref to the new PDFXrefEntry class.

+++ PDFRenderer/src/com/sun/pdfview/helper/XYRectFloat.java
@@ -28,13 +28,13 @@ public class XYRectFloat
 	//Only has what is necessary
 	
 	/** Height of rect. */
-	public float height;
+	public double height;
 	/** Width of rect. */
-	public float width;
+	public double width;
 	/** Horizontal position of top left point. */
-	public float x;
+	public double x;
 	/** Vertical position of top left point. */
-	public float y;
+	public double y;
 	
 	/**
 	 * Constructs a new XYRectFloat instance and sets all fields to 0.
@@ -51,7 +51,7 @@ public class XYRectFloat
 	 * @param width Distance in pixels to the right edge of the rectangle.
 	 * @param height Distance in pixels to the bottom edge of the rectangle.
 	 */
-	public XYRectFloat(float x, float y, float width, float height)
+	public XYRectFloat(double x, double y, double width, double height)
 	{
 		this.x = x;
 		this.y = y;
@@ -62,7 +62,7 @@ public class XYRectFloat
 	/**
 	 * Get the right edge of the rectangle.
 	 */
-	public float X2()
+	public double X2()
 	{
 		return x + width;
 	}
@@ -70,7 +70,7 @@ public class XYRectFloat
 	/**
 	 * Get the bottom edge of the rectangle.
 	 */
-	public float Y2()
+	public double Y2()
 	{
 		return y + height;
 	}
